+++ lbryum/commands.py
@@ -591,25 +591,91 @@ class Commands(object):
         return self.network.synchronous_get(('blockchain.transaction.broadcast', [str(tx)]))
 
     @command('w')
-    def history(self):
+    def get_auxilary_info_tx(self, txid):
+        info = {
+            'category': 'Unbound',
+            'name': "",
+            'claim_id': ""
+        }
+
+        tx = self.wallet.transactions[txid]
+        tx_outs = tx.outputs()
+
+        for nout, tx_out in enumerate(tx_outs):
+            if tx_out[0] & TYPE_SUPPORT:
+                info['category'] = 'Support'
+                claim_name, claim_id = tx_out[1][0]
+                info['name'] = claim_name
+                info['claim_id'] = encode_claim_id_hex(claim_id)
+            elif tx_out[0] & TYPE_UPDATE:
+                info['category'] = 'Update'
+                claim_name, claim_id, claim_value = tx_out[1][0]
+                info['name'] = claim_name
+                info['claim_id'] = encode_claim_id_hex(claim_id)
+            elif tx_out[0] & TYPE_CLAIM:
+                info['category'] = 'Claim'
+                claim_name, claim_value = tx_out[1][0]
+                info['name'] = claim_name
+                claim_id = claim_id_hash(rev_hex(tx.hash()).decode('hex'), nout)
+                claim_id = encode_claim_id_hex(claim_id)
+                info['claim_id'] = claim_id
+
+        return info
+
+    @staticmethod
+    def _get_tip_txns(name_claims):
+        tip_txns = set()
+
+        def is_tip(address):
+            for name_claim in name_claims:
+                if name_claim['address'] == address:
+                    if name_claim['category'] == "claim":
+                        return True
+
+            return False
+
+        for name_claim in name_claims:
+            if name_claim['category'] == "support":
+                tip = is_tip(name_claim['address'])
+            else:
+                tip = False
+
+            if tip:
+                tip_txns.add(name_claim['txid'])
+
+        return tip_txns
+
+    @command('w')
+    def history(self, include_tip_info=False):
         """Wallet history. Returns the transaction history of your wallet."""
         balance = 0
         out = []
+        if include_tip_info:
+            tip_txns = self._get_tip_txns(self.getnameclaims())
         for item in self.wallet.get_history():
             tx_hash, conf, value, timestamp, balance = item
+            aux_tx_info = self.get_auxilary_info_tx(tx_hash)
+
+            tip = False
+            if include_tip_info:
+                tip = True if tx_hash in tip_txns else False
+
             try:
                 time_str = datetime.datetime.fromtimestamp(timestamp).isoformat(' ')[:-3]
             except Exception:
                 time_str = "----"
-            label = self.wallet.get_label(tx_hash)
+
             out.append({
                 'txid': tx_hash,
                 'timestamp': timestamp,
                 'date': "%16s" % time_str,
-                'label': label,
                 'value': float(value) / float(COIN) if value is not None else None,
-                'confirmations': conf}
-            )
+                'confirmations': conf,
+                'type': aux_tx_info['category'],
+                'claim_id': aux_tx_info['claim_id'],
+                'is_tip': tip,
+                'claim_name': aux_tx_info['name']
+            })
         return out
 
     @command('w')
@@ -2030,7 +2096,8 @@ command_options = {
                           "do not check for an existing unspent claim before making a new one"),
     'revoke': (None, "--revoke", "if true, create a new signing key and revoke the old one"),
     'val': (None, '--value', 'claim value'),
-    'timeout': (None, '--timeout', 'timeout')
+    'timeout': (None, '--timeout', 'timeout'),
+    'include_tip_info': (None, "--include_tip_info", "if true, checks the tx list for tip txns(makes network call)")
 }